requirements.txt
tensorflow==2.7.0
pandas==1.3.4
seaborn==0.11.2
matplotlib==3.4.3
opencv-python==4.5.4.60
streamlit==1.2.0
pillow==8.4.0
numpy==1.20.3